// src/models/order-request.models.ts
import { Address } from "./address.models";
import { CartItem } from "./cart-item.models";
import { ShippingLine } from "./shipping-line.models";

export class OrderRequest {
    payment_method: string;
    payment_method_title: string;
    customer_id: string;
    set_paid: boolean;
    billing: Address;
    shipping: Address;
    line_items: Array<CartItem>;
    shipping_lines: Array<ShippingLine>;
    fee_lines: Array<{ name: string, total: string, tax_status: string }>;
}